Updated CategoriesFragment and ProductsLazyList

+++ app/src/main/java/com/vodovoz/app/domain/general/model/VodovozAction.kt
@@ -1,12 +1,14 @@
 package com.vodovoz.app.domain.general.model
 
 sealed interface VodovozAction {
-    data class Product(val id: Int) : VodovozAction
-    data class Products(val id: Int) : VodovozAction
-    data class Category(val id: Int) : VodovozAction
-    data class Promotion(val id: Int) : VodovozAction
-    data class Promotions(val categoryId: Int) : VodovozAction
-    data class Brand(val id: Int) : VodovozAction
+    data class Product(val id: Long) : VodovozAction
+    //TODO - update params: blockId, bannerId
+    data class Products(val id: Long) : VodovozAction
+    data class Category(val id: Long) : VodovozAction
+    data class Promotion(val id: Long) : VodovozAction
+    //TODO - update params: blockId, bannerId
+    data class Promotions(val categoryId: Long) : VodovozAction
+    data class Brand(val id: Long) : VodovozAction
     data class Url(val url: String) : VodovozAction
     data class UrlWithCookie(val url: String) : VodovozAction
 
@@ -16,7 +18,6 @@ sealed interface VodovozAction {
 
 enum class DataAllAction : VodovozAction {
     AllDiscount, AllNewProducts, AllPromotions, Delivery, Profile, WaterTracker, BuyCertificate, Unknown
-    //TODO - BuyCertificate
 }
